refactor: Simplify CalendarScreen structure by delegating to specialized components
- Refactor CalendarScreen.kt to act as a facade, delegating to specialized components for better maintainability and organization.
- Split functionality into dedicated files: CalendarApp.kt for app-level components, MainCalendarScreen.kt for main screen layout, and FooterComponents.kt for footer elements.
- Remove legacy UI logic to enhance clarity and streamline the component structure.
